refactor(commands): :art: Reorganize command view into different files
